requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pymupdf>=1.23.0
rank-bm25>=0.2.2
PyYAML>=6.0
Pillow>=10.0.0
pytesseract>=0.3.10
